feat: block remove list when not my likeable person

+++ src/main/java/com/ll/gramgram/boundedContext/likeablePerson/controller/LikeablePersonController.java
@@ -9,12 +9,14 @@ import jakarta.validation.Valid;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.security.Principal;
 import java.util.List;
@@ -65,6 +67,11 @@ public class LikeablePersonController {
     @GetMapping("/delete/{id}")
     public String delete(Principal principal, @PathVariable("id") Integer id) {
         LikeablePerson likeablePerson = this.likeablePersonService.getLikeablePerson(id);
+
+        if (!rq.getMember().getInstaMember().getUsername().equals(likeablePerson.getFromInstaMemberUsername())) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제권한이 없습니다");
+        }
+
         likeablePersonService.delete(likeablePerson);
 
         return "redirect:/";
